feat: 23 - refactor code
