+++ src/pages/registration/constants.ts
@@ -15,7 +15,7 @@ type RegistrationInputsProps = Omit<Options<'input'>, 'children' | 'parent' | 't
 export const REGISTRATION = {
   addressBlock: ['flex', 'flex-col', 'gap-1', 'border', 'border-[#252525]/50', 'p-2'],
   container: ['flex', 'justify-center', 'items-center', 'min-h-screen', 'p-3'],
-  errorsWrapper: ['w-full', 'max-w-[400px]', 'p-2' ],
+  errorsWrapper: ['w-full', 'max-w-[400px]', 'p-2'],
   form: ['flex', 'flex-col', 'w-full', 'items-center', 'gap-4'],
   inputsContainer: ['flex', 'gap-[10px]', 'w-full', 'mb-4', 'flex-wrap'],
   wrapper: [
